+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   `BackupManager::getEmailInstance()` method is also deprecated);
 * setting the `DatabaseBackup.mailSender` value of the configuration is deprecated (bootstrap checks that the value
   has not been set by the user);
+* the `send` option for the `ExportCommand` is deprecated. Will be removed in a future release;
 * added tests for php 8.4;
 * all chainable methods of X and Y now have the typehint for returning self. Updated descriptions;
 * updated `psalm` to `6.x`.